am,x86-nemu: modify parameter type to match the operand length of `out` instruction
* this will reduce unnecessary instructions to perform extensions in binary

+++ am/arch/x86-nemu/include/arch.h
@@ -29,11 +29,11 @@ static inline uint32_t inl(int port) {
   return data;
 }
 
-static inline void outb(int port, int data) {
+static inline void outb(int port, uint8_t data) {
   asm volatile("outb %%al, %%dx" : : "a"(data), "d"((uint16_t)port));
 }
 
-static inline void outl(int port, int data) {
+static inline void outl(int port, uint32_t data) {
   asm volatile("outl %%eax, %%dx" : : "a"(data), "d"((uint16_t)port));
 }
